% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_file.R
\name{chk_file}
\alias{chk_file}
\title{Helper function to choose the excel file that will provide the data}
\usage{
chk_file(path)
}
\arguments{
\item{path}{(Character) the path that the excel file can be read from or null in case the user wants to have a file choose window displayed.}
}
\value{
the path that the user chose
}
\description{
Helper function to choose the excel file that will provide the data
}
\author{
Agapios Panos <panosagapios@gmail.com>
}
