% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_export_path.R
\name{choose_export_path}
\alias{choose_export_path}
\title{Helper function for choosing the export path for the excel files}
\usage{
choose_export_path(export_data_path = NULL, suppressNA = TRUE)
}
\arguments{
\item{export_data_path}{(Character) (optional) (default: NULL (a popup message to choose dir will be displayed)) the path for excel file output.}

\item{suppressNA}{(Boolean) (optional) (default: TRUE) set this to FALSE if you want to know for which patients have NAs and for which variable. By default all NAs will be ignored so that the algorithm can distinguish between patients who meet the criterion and those who do not.}
}
\value{
a valid export path for the excel files.
}
\description{
Helper function for choosing the export path for the excel files
}
\author{
Agapios Panos <panosagapios@gmail.com>
}
