% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_crit.R
\name{eval_crit}
\alias{eval_crit}
\title{Evaluates the imported patients' data for the selected DDI criteria at once.}
\usage{
eval_crit(
  selected = "all",
  excel_path = NULL,
  exclude = NULL,
  excel_out = TRUE,
  single_excel = TRUE,
  export_data_path = NULL,
  suppressNA = TRUE,
  excel_sheet = 1,
  excel_col_data = "med_gen__decod",
  excel_col_pid = "usubjid",
  excel_col_daily_dosage = "daily_dosage",
  excel_col_med_unit = "med_unit",
  show_only_meet = FALSE,
  show_only_sum = FALSE
)
}
\arguments{
\item{selected}{(Character vector) (optional) (default: 'all') the selected criteria to export given in the form of c("DDI1, "DDI6"). Valid options 'all' and DDI1 to DDI69.}

\item{excel_path}{(Character) (optional) (default: NULL) the path that the excel file can be read from. If NULL a file choose window will be displayed so that you can choose the excel file.}

\item{exclude}{(Character) (optional) (default: NULL) a vector of criteria that you want to exclude. Example: c("DDI1, "DDI6"). Valid options DDI1 to DDI69.}

\item{excel_out}{(Boolean) (optional) (default: TRUE) output excel file with the evaluated data.}

\item{single_excel}{(Boolean) (optional) (default: TRUE) if true outputs only 1 excel file with multiple columns instead of multiple files (one for each criterion)}

\item{export_data_path}{(Character) (optional) (default: NULL (a popup message to choose dir will be displayed)) the path for excel file output.}

\item{suppressNA}{(Boolean) (optional) (default: TRUE) set this to FALSE if you want to know for which patients have NAs and for which variable. By default all NAs will be ignored so that the algorithm can distinguish between patients who meet the criterion and those who do not.}

\item{excel_sheet}{(Character) (optional) (default: 1) the number of excel worksheet that the patient id and the ATC codes are stored.}

\item{excel_col_data}{(Character) (optional) (default: 'med_gen__decod') the column name of the column that stores the ATC codes data in the provided excel file.}

\item{excel_col_pid}{(Character) (optional) (default: 'usubjid') the column name that specifies the patient id in the excel you provided for the data.}

\item{excel_col_daily_dosage}{(Character) (optional) (default: 'daily_dosage') the column name that specifies the daily dosage so that some DDI can take into account the dosage.}

\item{excel_col_med_unit}{(Character) (optional) (default: med_unit) the column name that specifies the medicine unit (e.g. mg or g) that is used to calculate the daily dosage.}

\item{show_only_meet}{(Boolean) (optional) (default: FALSE) set to TRUE if you want to have exported in the excel only the patients that meet the conditions for this criterion}

\item{show_only_sum}{(Boolean) (optional) (default: FALSE) set to TRUE if you want to show only the number of number of criteria that are met for each patient and a boolean value of 0 if no criterion is met and 1 if at least one is met.}
}
\description{
Evaluates the imported patients' data for the selected DDI criteria at once.
}
\author{
Agapios Panos <panosagapios@gmail.com>
}
