% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bool_exp.R
\name{get_bool_exp}
\alias{get_bool_exp}
\title{Helper function to get the correspoding to that criterion boolean expressions for the grepl function}
\usage{
get_bool_exp(criterion)
}
\arguments{
\item{criterion}{(Character) The specific criterion for which you need the boolean expressions.}
}
\value{
Returns a list of 2 boolean expressions to pass as arguments to the grepl function.
}
\description{
Helper function to get the correspoding to that criterion boolean expressions for the grepl function
}
