% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel_data.R
\name{import_excel_data}
\alias{import_excel_data}
\title{Imports patient data from excel file.}
\usage{
import_excel_data(current_data, path, worksheet, var_col,
  ignore_na = FALSE, include_missing = FALSE,
  excel_col_pid = "usubjid")
}
\arguments{
\item{current_data}{(List) the data that was imported by another call of import_excel_data so that this call will append the new data to the old ones.}

\item{path}{(Character) the path that the excel file can be read from.}

\item{worksheet}{(Character) the worksheet that the desired variable is stored in.}

\item{var_col}{(Character) the column that stores the desired variable}

\item{ignore_na}{(Boolean) (optional) (default: FALSE) a boolean variable that indicates whether to ignore any N/As found in excel (if set to TRUE) or to handle them as missing data (if set to FALSE).}

\item{include_missing}{(Boolean) (optional) (default: FALSE) a boolean that indicated whether to include the missing values as NAs in the data to achieve one to one correspondance in the data if necessary.}

\item{excel_col_pid}{(Character) the column name that specifies the patient id in the excel you provided for the data.}
}
\value{
list of [[1]] patient data for the specified variable. [[2]] missing patient id and variable name as specified in excel that we used to import data.
}
\description{
Imports patient data from excel file.
}
\author{
Agapios Panos <panosagapios@gmail.com>
}
