% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_crit.R
\name{validate_crit}
\alias{validate_crit}
\title{Validation for the selected and exclude argument for the names of criteria specified}
\usage{
validate_crit(selected, exclude)
}
\arguments{
\item{selected}{a vector of the selected criteria for which you want to perform the patient data evaluation.}

\item{exclude}{a vector of the criteria that you want to exclude from the evaluation.}
}
\description{
Validation for the selected and exclude argument for the names of criteria specified
}
