% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi.R
\name{ddc}
\alias{ddc}
\title{Data Defect Correlation}
\usage{
ddc(data = df, N = "N", n = "n", mu, muhat, cv = NULL)
}
\arguments{
\item{data}{dataframe}

\item{N}{variable name (char) for population size. Default assumes \code{"N"}}

\item{n}{variable name (char) for sample size. Default assumes \code{"n"}}

\item{mu}{variable name (char) for population quantity of interest}

\item{muhat}{variable name (char) for sample estimate}
}
\description{
The Data Defect Correlation (d.d.c.) is simply the square root of d.d.i. d.d.c.
is the correlation between response and group memembership. Currently both variables
are binary. Squaring the d.d.c. is more useful for characterizing the asymptotics of
MSE.
}
\examples{
data(g2016)
ddcs <- ddc(g2016, N = "tot_votes", n = "cces_n_vv", mu = "pct_djt_voters", muhat = "cces_pct_djt_vv")
tibble(st = g2016$st, ddc = ddcs)


}
\references{
Meng, Xiao-Li. Statistical paradises and paradoxes in big data (I):
Law of large populations, big data paradox, and the 2016 US presidential election.
Ann. Appl. Stat. 12 (2018), no. 2, 685--726. \url{http://doi.org/10.1214/18-AOAS1161SF}
}
